% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siftgeojson-package.R
\docType{package}
\name{siftgeojson-package}
\alias{siftgeojson}
\alias{siftgeojson-package}
\title{sift geojson}
\description{
sift geojson
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

